refactor($validator): 패키지 위치 수정 (루트->도메인)
